+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - CASMCMS-9413: Add explicit annotations for S3 methods
   - CASMCMS-9414: Create helper types for sessions and component type annotation
   - CASMCMS-9415: Refactor `dbwrapper` to resolve type annotation issues
+  - CASMCMS-9416: Resolve type annotation issues in `server/options.py`
 
 ### Dependencies
 - Updated to openapi-generator-cli v7.13.0